server/requirements.txt
django==4.2.2
djangorestframework==3.14.0
django-filter==23.2
Pillow==9.5.0
django-cors-headers==4.1.0
django_q2==1.5.3
weasyprint==58.1
libsass==0.22.0
django-modeltranslation==0.18.10
markdown==3.4.3
bleach==6.0.0
pygments==2.15.1
drf-spectacular==0.26.2
matplotlib==3.7.1
cvss==2.6
jinja2==3.1.2
argon2-cffi==21.3.0